requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
beautifulsoup4==4.14.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
cssselect2==0.8.0
distro==1.9.0
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-htmx==1.26.0
django-taggit==6.1.0
djangorestframework==3.16.1
fonttools==4.60.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
lxml==6.0.2
Markdown==3.6
openai==2.6.0
packaging==25.0
pillow==11.3.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
zopfli==0.2.3.post1
